+++ cmd/openshift-install/gather.go
@@ -15,11 +15,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/rest"
 
-	configv1 "github.com/openshift/api/config/v1"
-	configclient "github.com/openshift/client-go/config/clientset/versioned"
 	"github.com/openshift/installer/cmd/openshift-install/command"
 	"github.com/openshift/installer/pkg/asset/installconfig"
 	assetstore "github.com/openshift/installer/pkg/asset/store"
@@ -226,37 +222,6 @@ func pullLogsFromBootstrap(gatherID string, bootstrap string, port int, masters 
 	}
 
 	return clusterLogBundlePath, nil
-}
-
-func logClusterOperatorConditions(ctx context.Context, config *rest.Config) error {
-	client, err := configclient.NewForConfig(config)
-	if err != nil {
-		return fmt.Errorf("creating a config client: %w", err)
-	}
-
-	operators, err := client.ConfigV1().ClusterOperators().List(ctx, metav1.ListOptions{})
-	if err != nil {
-		return fmt.Errorf("listing ClusterOperator objects: %w", err)
-	}
-
-	for _, operator := range operators.Items {
-		for _, condition := range operator.Status.Conditions {
-			if condition.Type == configv1.OperatorUpgradeable {
-				continue
-			} else if condition.Type == configv1.OperatorAvailable && condition.Status == configv1.ConditionTrue {
-				continue
-			} else if (condition.Type == configv1.OperatorDegraded || condition.Type == configv1.OperatorProgressing) && condition.Status == configv1.ConditionFalse {
-				continue
-			}
-			if condition.Type == configv1.OperatorAvailable || condition.Type == configv1.OperatorDegraded {
-				logrus.Errorf("Cluster operator %s %s is %s with %s: %s", operator.ObjectMeta.Name, condition.Type, condition.Status, condition.Reason, condition.Message)
-			} else {
-				logrus.Infof("Cluster operator %s %s is %s with %s: %s", operator.ObjectMeta.Name, condition.Type, condition.Status, condition.Reason, condition.Message)
-			}
-		}
-	}
-
-	return nil
 }
 
 func gatherCAPIArtifacts(directory, gatherID string) (string, error) {
