+++ cmd/openshift-install/waitfor.go
@@ -10,6 +10,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/openshift/installer/cmd/openshift-install/command"
+	assetstore "github.com/openshift/installer/pkg/asset/store"
 	timer "github.com/openshift/installer/pkg/metrics/timer"
 )
 
@@ -49,7 +50,7 @@ func newWaitForBootstrapCompleteCmd() *cobra.Command {
 			}
 			timer.StartTimer("Bootstrap Complete")
 			if err := waitForBootstrapComplete(ctx, config); err != nil {
-				if err2 := logClusterOperatorConditions(ctx, config); err2 != nil {
+				if err2 := command.LogClusterOperatorConditions(ctx, config); err2 != nil {
 					logrus.Error("Attempted to gather ClusterOperator status after wait failure: ", err2)
 				}
 
@@ -57,7 +58,7 @@ func newWaitForBootstrapCompleteCmd() *cobra.Command {
 				logrus.Info("openshift-install gather bootstrap --help")
 				logrus.Error("Bootstrap failed to complete: ", err.Unwrap())
 				logrus.Error(err.Error())
-				logrus.Exit(exitCodeBootstrapFailed)
+				logrus.Exit(command.ExitCodeBootstrapFailed)
 			}
 
 			logrus.Info("It is now safe to remove the bootstrap resources")
@@ -85,14 +86,20 @@ func newWaitForInstallCompleteCmd() *cobra.Command {
 				logrus.Fatal(errors.Wrap(err, "loading kubeconfig"))
 			}
 
-			err = waitForInstallComplete(ctx, config, command.RootOpts.Dir)
+			assetStore, err := assetstore.NewStore(command.RootOpts.Dir)
 			if err != nil {
-				if err2 := logClusterOperatorConditions(ctx, config); err2 != nil {
+				logrus.Error(err)
+				logrus.Exit(command.ExitCodeInstallFailed)
+			}
+
+			err = command.WaitForInstallComplete(ctx, config, assetStore)
+			if err != nil {
+				if err2 := command.LogClusterOperatorConditions(ctx, config); err2 != nil {
 					logrus.Error("Attempted to gather ClusterOperator status after wait failure: ", err2)
 				}
-				logTroubleshootingLink()
+				command.LogTroubleshootingLink()
 				logrus.Error(err)
-				logrus.Exit(exitCodeInstallFailed)
+				logrus.Exit(command.ExitCodeInstallFailed)
 			}
 			timer.StopTimer(timer.TotalTimeElapsed)
 			timer.LogSummary()
